Add defensive programming for product manipulation and user profile manipulation

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -89,6 +90,7 @@ def product_detail(request, product_id):
     return render(request, url, context)
 
 
+@login_required
 def add_product(request):
     """ 
     Add a product to the store.
@@ -101,6 +103,10 @@ def add_product(request):
     new_arrivals = Product.objects.filter(new_product=True)
     brand_item = Product.objects.filter(brand_item=True)
     five_star_product = Product.objects.filter(rating__gte=5)
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners and trusted partners can do that.')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -128,9 +134,17 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
-    """ Edit a product in the store """
+    """ 
+    Edit a product in the store 
+    """
     product = get_object_or_404(Product, pk=product_id)
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners and trusted partners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
@@ -152,10 +166,15 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ 
     Delete a product from the store 
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners and trusted partners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
